requirements.txt
Flask
requests
# release version (0.7.5) has a bug and won't run in py3
locustio==0.8a2
